Integ tests changes  (#338)
This PR adds support for running integration tests for index-management plugin as part of the CICD framework. It also provides a template implementation for test_manifest, which is a configuration file that defines test support matrix for components of a bundle. This file is schema-version'ed and subject to change in future as the template evolves.  Currently, the dependency management for components is done in a legacy way by building dependencies for job-scheduler and notifications. This logic will be replaced soon with DependencyInstaller #283 . 

* Integ tests changes - working integtests, code refactoring, dependency management, test manifest

Signed-off-by: Himanshu Setia <setiah@amazon.com>

* Addressing PR comments

Signed-off-by: Himanshu Setia <setiah@amazon.com>

* Adding licensing info to file

Signed-off-by: Himanshu Setia <setiah@amazon.com>

* correcting import order with pipenv isort

Signed-off-by: Himanshu Setia <setiah@amazon.com>

* resolving merge conflicts

Signed-off-by: Himanshu Setia <setiah@amazon.com>

* Removed placeholder bundle and build manifest files

Signed-off-by: Himanshu Setia <setiah@amazon.com>

* Removing unwanted integtest code from test.py

Signed-off-by: Himanshu Setia <setiah@amazon.com>

* pipenv isort fixes

Signed-off-by: Himanshu Setia <setiah@amazon.com>

* pipenv flake fixes

Signed-off-by: Himanshu Setia <setiah@amazon.com>