[Build] Add OpenSearch Dashboards to Jenkinsfile (#717)
* [Build] create Dashboards Jenkinsfile and move OpenSearch Jenkinsfile

Creates OpenSearch Dashboards Jenkinsfile basically copying what was
in the existing OpenSearch Jenkinsfile. But skipping snapshots since
it was not doing anything with snapshots and publishing to S3
bucket paths of `/builds/dashboards/` and `/bundles/dashboards/` which
has not been verified yet. It was done this way, while keeping OpenSearch
to build to `/builds` and `/bundles`, to reduce the impact of adding
OpenSearch Dashboards.

For consistency, moved the root lvl Jenkinsfile which was only  building
OpenSearch to `jenkins/opensearch`.

Issue resolved:
https://github.com/opensearch-project/opensearch-build/issues/620

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>

* [Build] remove Java arg, x86 to x64, s3 bucket rename

Handling the feedback without any question:

* Remove JAVA arg since it is unneeded for OpenSearch Dashboards.
* Change from x86 to x64
* Change S3 bucket from dashboards to opensearch-dashboards

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>

* [Build] rename folder and update build id for dashboards

Update from feedback with potential to require more discussion.
* Rename folder from jenkins/opensearch_dashboards to jenkins/opensearch-dashboards
* Build ID from OPENSEARCH_BUILD_ID to OPENSEARCH_DASHBOARDS_BUILD_ID

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>